+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
  * `LazyBinding` does not cache resources anymore in case the repository 
    contents changed since the last call
+ * removed `NoQueryMatchesException`
 
 * 1.0.0-beta3 (2015-03-19)